LayoutTests/fast/viewport/watchos/viewport-with-system-minimum-layout-margins.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true horizontalSystemMinimumLayoutMargin=12 ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<meta name="disabled-adaptations" content="watch">
<head>
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<style>
body, html {
    margin: 0;
    width: 100%;
    height: 100%;
}
</style>
<script>
jsTestIsAsync = true;

async function runTest() {
    description("This test verifies that non-zero horizontal system minimum layout margins affect the width of the viewport.");

    if (!window.testRunner)
        return;

    await UIHelper.ensureVisibleContentRectUpdate();
    shouldBe("innerWidth", "160");
    finishJSTest();
}
</script>
</head>
<body onload="runTest()"></body>
</html>
