LayoutTests/fast/viewport/watchos/viewport-with-system-minimum-layout-margins-expected.txt

This test verifies that non-zero horizontal system minimum layout margins affect the width of the viewport.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS innerWidth is 132
PASS successfullyParsed is true

TEST COMPLETE
